pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gabber-engine"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.12.2"

# Core dependencies needed by both modes
dependencies = [
    "livekit~=1.0.13",
    "livekit-api>=1.0.5",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = ["ruff>=0.12.5", "watchfiles>=1.1.0"]

[tool.ruff]
# Set the target Python version (modern Python, aligns with projects like FastAPI and Pandas)
target-version = "py312"

# Set line length to match Black's default (88 characters), common in many projects
line-length = 88
indent-width = 4

# Include commonly ignored directories to avoid linting irrelevant files
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

src = ["gabber"]

# Enable a broad set of lint rules for comprehensive checks
[tool.ruff.lint]
fixable = ["ALL"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Use Google-style docstring conventions (common in projects like Pandas)
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.hatch.build.targets.wheel]
packages = ["gabber"]
